Centralize documented param name extraction
