+++ src/plugin/src/ast-transforms/condense-logical-expressions.js
@@ -1,18 +1,12 @@
 import {
-    getCommentArray,
-    hasComment as sharedHasComment,
-    isDocCommentLine,
-    normalizeHasCommentHelpers
+    normalizeHasCommentHelpers,
+    getDocCommentManager
 } from "../comments/index.js";
-import {
-    cloneLocation,
-    getNodeStartIndex
-} from "../../../shared/ast-locations.js";
+import { cloneLocation } from "../../../shared/ast-locations.js";
 import { isNonEmptyArray } from "../../../shared/array-utils.js";
 import { getBodyStatements, isNode } from "../../../shared/ast-node-helpers.js";
 import {
     isNonEmptyString,
-    isNonEmptyTrimmedString,
     toNormalizedLowerCaseString
 } from "../../../shared/string-utils.js";
 import { getOrCreateMapEntry } from "../../../shared/object-utils.js";
@@ -25,9 +19,7 @@ const BOOLEAN_NODE_TYPES = Object.freeze({
     OR: "OR"
 });
 
-const DEFAULT_HELPERS = Object.freeze({
-    hasComment: sharedHasComment
-});
+const DEFAULT_HELPERS = Object.freeze(normalizeHasCommentHelpers());
 
 const LOGICAL_OPERATORS = new Set(["and", "&&", "or", "||"]);
 const COMPARISON_OPERATORS = new Set(["==", "!=", "<>", "<=", ">=", "<", ">"]);
@@ -52,104 +44,21 @@ export function condenseLogicalExpressions(ast, helpers) {
         return ast;
     }
 
-    normalizeDocCommentWhitespace(ast);
+    const docCommentManager = getDocCommentManager(ast);
     const normalizedHelpers = normalizeHasCommentHelpers(helpers);
     const context = {
         ast,
         helpers: normalizedHelpers,
         docUpdates: new Map(),
-        commentGroups: null,
+        docCommentManager,
         expressionSignatures: new Map()
     };
     activeTransformationContext = context;
     visit(ast, normalizedHelpers, null);
-    applyDocCommentUpdates(context);
+    docCommentManager.applyUpdates(context.docUpdates);
     removeDuplicateCondensedFunctions(context);
     activeTransformationContext = null;
     return ast;
-}
-
-function normalizeDocCommentWhitespace(ast) {
-    const comments = getCommentArray(ast);
-
-    if (comments.length === 0) {
-        return;
-    }
-
-    for (const comment of comments) {
-        if (
-            comment?.type === "CommentLine" &&
-            typeof comment.leadingWS === "string" &&
-            /(?:\r\n|\r|\n|\u2028|\u2029)\s*(?:\r\n|\r|\n|\u2028|\u2029)/.test(
-                comment.leadingWS
-            )
-        ) {
-            comment.leadingWS = "\n";
-        }
-    }
-}
-
-function extractDescriptionContent(value) {
-    if (typeof value !== "string") {
-        return "";
-    }
-
-    return value.replace(/^\s*\/\s*@description\s*/i, "").trim();
-}
-
-function buildUpdatedDescription(existing, expression) {
-    if (!expression) {
-        return existing ?? "";
-    }
-
-    const normalizedExpression = expression.trim();
-
-    if (!isNonEmptyTrimmedString(existing)) {
-        return `Simplified: ${normalizedExpression}`;
-    }
-
-    const trimmed = existing.trim();
-    const lowered = trimmed.toLowerCase();
-
-    if (lowered.includes("original multi-branch")) {
-        return existing ?? "";
-    }
-
-    if (lowered.includes("original") || lowered.includes("multi-clause")) {
-        return `Simplified: ${normalizedExpression}`;
-    }
-
-    if (lowered.includes("simplified")) {
-        const colonIndex = trimmed.indexOf(":");
-        if (colonIndex !== -1) {
-            const prefix = trimmed.slice(0, colonIndex + 1);
-            return `${prefix} ${normalizedExpression}`;
-        }
-        return `Simplified: ${normalizedExpression}`;
-    }
-
-    if (lowered.includes("guard extraction")) {
-        return existing ?? "";
-    }
-
-    if (trimmed.includes("==")) {
-        const equalityIndex = trimmed.indexOf("==");
-        const prefix = trimmed.slice(0, equalityIndex + 2).trimEnd();
-        return `${prefix} ${normalizedExpression}`;
-    }
-
-    const mentionsReturn = /\breturn\b/.test(lowered);
-    const mentionsBranching =
-        /\bif\b/.test(lowered) || /\belse\b/.test(lowered);
-
-    if (mentionsReturn && mentionsBranching) {
-        return existing ?? "";
-    }
-
-    const withoutPeriod = trimmed.replace(/\.?\s*$/, "");
-    const needsSemicolon = mentionsReturn;
-    const separator = needsSemicolon ? "; ==" : " ==";
-    return `${withoutPeriod}${separator} ${normalizedExpression}`;
 }
 
 function isBooleanBranchExpression(node, allowValueLiterals = false) {
@@ -230,18 +139,12 @@ function isBooleanBranchExpression(node, allowValueLiterals = false) {
     }
 }
 
-function ensureCommentGroups(context) {
-    if (!context.commentGroups) {
-        context.commentGroups = mapDocCommentsToFunctions(context.ast);
-    }
-    return context.commentGroups;
-}
-
 function removeDuplicateCondensedFunctions(context) {
     if (!context || !Array.isArray(context.ast?.body)) {
         return;
     }
 
+    const docCommentManager = context.docCommentManager;
     const signatureToFunctions = new Map();
     for (const [fn, signature] of context.expressionSignatures.entries()) {
         if (!signature) {
@@ -257,7 +160,6 @@ function removeDuplicateCondensedFunctions(context) {
         return;
     }
 
-    const commentGroups = ensureCommentGroups(context);
     const toRemove = new Set();
 
     for (const functions of signatureToFunctions.values()) {
@@ -270,7 +172,7 @@ function removeDuplicateCondensedFunctions(context) {
             const update = context.docUpdates.get(fn);
             const hasDocComment =
                 update?.hasDocComment ||
-                (commentGroups.get(fn)?.length ?? 0) > 0;
+                (docCommentManager?.hasDocComment(fn) ?? false);
             if (hasDocComment && !keeper) {
                 keeper = fn;
             }
@@ -285,7 +187,7 @@ function removeDuplicateCondensedFunctions(context) {
                 const update = context.docUpdates.get(fn);
                 const hasDocComment =
                     update?.hasDocComment ||
-                    (commentGroups.get(fn)?.length ?? 0) > 0;
+                    (docCommentManager?.hasDocComment(fn) ?? false);
                 if (!hasDocComment) {
                     toRemove.add(fn);
                 }
@@ -298,176 +200,6 @@ function removeDuplicateCondensedFunctions(context) {
     }
 
     context.ast.body = context.ast.body.filter((node) => !toRemove.has(node));
-}
-
-function mapDocCommentsToFunctions(ast) {
-    const functions = collectFunctionNodes(ast).sort((a, b) => {
-        const aStart = getNodeStartIndex(a) ?? 0;
-        const bStart = getNodeStartIndex(b) ?? 0;
-        return aStart - bStart;
-    });
-
-    const groups = new Map();
-    for (const fn of functions) {
-        groups.set(fn, []);
-    }
-
-    const astComments = getCommentArray(ast);
-
-    if (astComments.length === 0 || functions.length === 0) {
-        return groups;
-    }
-
-    let functionIndex = 0;
-    for (const comment of astComments) {
-        if (!isDocCommentLine(comment)) {
-            continue;
-        }
-
-        const commentIndex = comment?.start?.index;
-        if (typeof commentIndex !== "number") {
-            continue;
-        }
-
-        while (functionIndex < functions.length) {
-            const targetStart = getNodeStartIndex(functions[functionIndex]);
-            if (typeof targetStart !== "number" || targetStart > commentIndex) {
-                break;
-            }
-            functionIndex += 1;
-        }
-
-        if (functionIndex >= functions.length) {
-            break;
-        }
-
-        const targetFunction = functions[functionIndex];
-        const bucket = groups.get(targetFunction);
-        if (bucket) {
-            bucket.push(comment);
-        }
-    }
-
-    return groups;
-}
-
-function applyDocCommentUpdates(context) {
-    if (!context || context.docUpdates.size === 0) {
-        return;
-    }
-
-    const commentGroups = ensureCommentGroups(context);
-
-    for (const [fn, update] of context.docUpdates.entries()) {
-        if (!update || !isNonEmptyTrimmedString(update.expression)) {
-            continue;
-        }
-
-        const comments = commentGroups.get(fn);
-        if (!comments || comments.length === 0) {
-            continue;
-        }
-
-        const descriptionComment = comments.find(
-            (comment) =>
-                typeof comment?.value === "string" &&
-                /@description\b/i.test(comment.value)
-        );
-
-        if (!descriptionComment) {
-            continue;
-        }
-
-        let updatedDescription = buildUpdatedDescription(
-            update.description,
-            update.expression
-        );
-
-        if (!isNonEmptyTrimmedString(updatedDescription)) {
-            continue;
-        }
-
-        const originalDescription =
-            typeof update.description === "string"
-                ? update.description.trim()
-                : "";
-
-        if (
-            originalDescription.endsWith(".") &&
-            !/[.!?]$/.test(updatedDescription)
-        ) {
-            updatedDescription = `${updatedDescription}.`;
-        }
-
-        const existingDescription = extractDescriptionContent(
-            descriptionComment.value
-        );
-
-        if (existingDescription === updatedDescription) {
-            continue;
-        }
-
-        const prefixMatch = descriptionComment.value.match(
-            /^(\s*\/\s*@description\s*)/i
-        );
-        const prefix = prefixMatch ? prefixMatch[1] : "/ @description ";
-
-        descriptionComment.value = `${prefix}${updatedDescription}`;
-    }
-}
-
-function collectFunctionNodes(ast) {
-    const functions = [];
-
-    function traverse(node) {
-        if (!isNode(node)) {
-            return;
-        }
-
-        if (node.type === "FunctionDeclaration") {
-            functions.push(node);
-        }
-
-        for (const [key, value] of Object.entries(node)) {
-            if (key === "start" || key === "end" || key === "comments") {
-                continue;
-            }
-
-            if (Array.isArray(value)) {
-                for (const child of value) {
-                    traverse(child);
-                }
-            } else if (isNode(value)) {
-                traverse(value);
-            }
-        }
-    }
-
-    traverse(ast);
-    return functions;
-}
-
-function extractFunctionDescription(ast, functionNode) {
-    if (!activeTransformationContext) {
-        return null;
-    }
-
-    const groups = ensureCommentGroups(activeTransformationContext);
-    const comments = groups.get(functionNode);
-    if (!comments) {
-        return null;
-    }
-
-    for (const comment of comments) {
-        if (
-            typeof comment.value === "string" &&
-            comment.value.includes("@description")
-        ) {
-            return extractDescriptionContent(comment.value);
-        }
-    }
-
-    return null;
 }
 
 function renderExpressionForDocComment(expressionAst) {
@@ -774,10 +506,11 @@ function tryCondenseIfStatement(
         activeTransformationContext
     ) {
         const docString = renderExpressionForDocComment(argumentAst);
-        const description = extractFunctionDescription(
-            activeTransformationContext.ast,
-            parentNode
-        );
+        const docCommentManager =
+            activeTransformationContext.docCommentManager;
+        const description = docCommentManager
+            ? docCommentManager.extractDescription(parentNode)
+            : null;
 
         if (docString) {
             activeTransformationContext.docUpdates.set(parentNode, {
