+++ src/plugin/src/ast-transforms/convert-manual-math.js
@@ -1,7 +1,4 @@
-import {
-    hasComment as sharedHasComment,
-    normalizeHasCommentHelpers
-} from "../comments/index.js";
+import { normalizeHasCommentHelpers } from "../comments/index.js";
 import {
     cloneLocation,
     getNodeEndIndex,
@@ -9,9 +6,7 @@ import {
 } from "../../../shared/ast-locations.js";
 import { getCallExpressionArguments } from "../../../shared/ast-node-helpers.js";
 
-const DEFAULT_HELPERS = Object.freeze({
-    hasComment: sharedHasComment
-});
+const DEFAULT_HELPERS = Object.freeze(normalizeHasCommentHelpers());
 
 const BINARY_EXPRESSION = "BinaryExpression";
 const CALL_EXPRESSION = "CallExpression";
